# packages/contracts/openapi/k-admission.v1.yaml
openapi: 3.1.0
info:
  title: K-Admission University Data Plane API
  version: 1.4.0
  description: >
    Reference contract for a university-local K-Admission data plane.
    The university DB is the source of truth; central sync is not in the applicant critical path.


    1.2.0 aligns the contract with the repository implementation (spec discrepancy register
    D-7, D-16, D-17, D-19, D-20, D-22, D-24, D-26, D-32, D-34, D-35, D-37, D-38, D-39, D-40).
    Paths marked with `x-kadmission-register` were added by the implementation first.


    1.3.0 adds the common `429` response (`RateLimited`, `Retry-After`) to applicant operations (D-51).


    1.4.0 connects what the implementation had but the contract did not describe (D-55 to D-61):
    the application state machine is driven by validate / payment / finalize (READY, PAYMENT_PENDING,
    PAID are now real states and editing stops at PAYMENT_PENDING); one live payment per application
    (`200` resumes the open checkout, `409 PAYMENT_IN_PROGRESS` refuses a second payment); the form
    schema carries `profileFields` and `documents` so the client draws everything from config; the
    receipt is audited; server clock offset against the DB is reported (§A9); the applicant-facing
    common profile API on the central host; the active config includes its body and config drafts
    are linted (§A5); universities send the §04 heartbeat and the dashboard shows which university
...


    Common rules:
...
servers:
...
    variables:
...
paths:
...
  /admin/v1/config/active:
    get:
      tags: [Admin]
      operationId: getActiveConfig
      x-kadmission-register: D-59
      description: >
        Includes the configuration body (`config`) — a new draft usually starts from the active one.
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/CycleIdQuery'
      responses:
        '200':
          description: Active configuration
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ConfigVersion'
                  - type: object
                    properties:
                      config: { type: object, additionalProperties: true }
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          description: No configuration is active for this cycle
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /admin/v1/config/versions:
    get:
      tags: [Admin]
      operationId: listConfigVersions
      x-kadmission-register: D-35
      summary: Approval inbox — versions without their body (newest first, at most 50)
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/CycleIdQuery'
      responses:
        '200':
          description: Config versions
          content:
            application/json:
              schema:
                type: object
                required: [versions]
                properties:
                  versions:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ConfigVersion'
                        - type: object
                          required: [createdAt]
                          properties:
                            createdAt: { type: string, format: date-time }
    post:
      tags: [Admin]
      operationId: createConfigVersion
      x-kadmission-register: D-59
      description: >
        The draft is linted with the same engine the runtime uses (§A5 Config Linter): a form JSON
        Schema that does not compile, `required` naming an undefined field, or malformed document
        lists refuse the draft with 400 — applied, they would stop every save and payment of that
        admission type. Softer findings are returned as `warnings` in the diff.
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateConfigVersionRequest' }
      responses:
        '201':
          description: Draft config version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigVersion' }
        '400':
          $ref: '#/components/responses/Problem'
  /admin/v1/config/versions/{configId}/diff:
    get:
      tags: [Admin]
      operationId: getConfigDiff
      x-kadmission-register: T-M3-02
      summary: What this version changes against the current ACTIVE one
      description: >
        Destructive changes are listed separately. The returned `digest` must be sent back when
        approving — an approval means "I agree to this change", not "I agree to this id".
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/ConfigId'
      responses:
        '200':
          description: Diff
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigDiff' }
  /admin/v1/config/versions/{configId}/approve:
    post:
      tags: [Admin]
      operationId: approveConfigVersion
      description: >
        Two distinct approvers, neither of them the author. `acknowledgedDiffDigest` must equal
        the current diff digest; if the draft or the ACTIVE baseline changed after review, 409.
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/ConfigId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [acknowledgedDiffDigest]
              properties:
                acknowledgedDiffDigest: { type: string }
      responses:
        '200':
          description: Approval recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ConfigVersion'
                  - type: object
                    required: [remainingApprovals]
                    properties:
                      remainingApprovals: { type: integer, minimum: 0, maximum: 2 }
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /admin/v1/config/versions/{configId}/activate:
    post:
      tags: [Admin]
      operationId: activateConfigVersion
      description: >
        Activates after two-person approval; the previous ACTIVE version is RETIRED in the same
        transaction. Refused (403) inside the pre-deadline freeze window or without two distinct non-author approvers. Every activation is recorded
        as a signed, append-only activation record.
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/ConfigId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateRequest' }
      responses:
        '200':
          description: Activated after two-person approval
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ConfigVersion'
                  - type: object
                    required: [activation]
                    properties:
                      activation: { $ref: '#/components/schemas/ActivationRecord' }
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /admin/v1/config/versions/{configId}/rollback:
    post:
      tags: [Admin]
      operationId: rollbackConfigVersion
      x-kadmission-register: T-M3-02
      summary: Re-activate a version that was ACTIVE before
      description: >
        Does not create a new version: the target row already carries two real approvals.
        Only versions that were actually activated before can be targets. Not blocked by the
        freeze window — freezing stops new changes, not recovery.
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/ConfigId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 2 }
      responses:
        '200':
          description: Restored
          content:
            application/json:
              schema:
                type: object
                required: [restored, retired, activation]
                properties:
                  restored: { $ref: '#/components/schemas/ConfigVersion' }
                  retired: { type: [string, 'null'], description: Version that was ACTIVE before the rollback }
                  activation: { $ref: '#/components/schemas/ActivationRecord' }
        '400':
          $ref: '#/components/responses/Problem'
  /admin/v1/deadline-policies:
    get:
      tags: [Admin]
      operationId: listDeadlinePolicies
      summary: Deadline policy history — who changed the deadline and when
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/CycleIdQuery'
      responses:
        '200':
          description: Policies
          content:
            application/json:
              schema:
                type: object
                required: [policies]
                properties:
                  policies:
                    type: array
                    items: { $ref: '#/components/schemas/DeadlinePolicy' }
    post:
      tags: [Admin]
      operationId: createDeadlinePolicy
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDeadlinePolicyRequest' }
      responses:
        '201':
          description: Deadline policy draft
          content:
            application/json:
              schema:
                type: object
                required: [policyId, policyHash]
                properties:
                  policyId: { type: string, format: uuid }
                  policyHash: { type: string }
        '400':
          $ref: '#/components/responses/Problem'
  /admin/v1/deadline-policies/extensions:
    post:
      tags: [Admin]
      operationId: createDeadlineExtension
      x-kadmission-register: D-35
      summary: Draft a deadline extension from the policy currently in force
      description: >
        Requires a reason and the admissions office decision document number — the platform
        records and executes the office's decision, it does not decide. Then the normal
        two-person approval applies. Not blocked by the freeze window. If the base policy is no
        longer in force at activation time, activation is refused (409).
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cycleId, deadlineAt, reason, decisionRef]
              properties:
                cycleId: { type: string, format: uuid }
                deadlineAt: { type: string, format: date-time }
                reason: { type: string, minLength: 5 }
                decisionRef: { type: string, minLength: 1 }
      responses:
        '201':
          description: Extension draft
          content:
            application/json:
              schema:
                type: object
                required: [policyId, policyHash, version, extendsVersion]
                properties:
                  policyId: { type: string, format: uuid }
                  policyHash: { type: string }
                  version: { type: string }
                  extendsVersion: { type: string }
        '400':
          $ref: '#/components/responses/Problem'
  /admin/v1/deadline-policies/{policyId}/approve:
    post:
      tags: [Admin]
      operationId: approveDeadlinePolicy
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/PolicyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Approval recorded
          content:
            application/json:
              schema:
                type: object
                required: [policyId, complete, remainingApprovals]
                properties:
                  policyId: { type: string, format: uuid }
                  complete: { type: boolean }
                  remainingApprovals: { type: integer, minimum: 0, maximum: 1 }
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          description: Another approval landed first (conditional update); reload and retry
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /admin/v1/deadline-policies/{policyId}/activate:
    post:
      tags: [Admin]
      operationId: activateDeadlinePolicy
      x-kadmission-register: D-22
      summary: Activate an approved deadline policy (optionally scheduled)
      description: >
        Symmetric with config activation. A policy that was applied once is never re-applied
        (no back door to an old deadline), and a deadline already in the past at activation time
        is refused (no retroactive closing). Recorded as a signed activation record.
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/PolicyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActivateRequest' }
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema:
                type: object
                required: [activatedAt, activation]
                properties:
                  activatedAt: { type: string, format: date-time }
                  activation: { $ref: '#/components/schemas/ActivationRecord' }
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /admin/v1/activations:
    get:
      tags: [Admin]
      operationId: listActivations
      x-kadmission-register: D-35
      summary: Signed activation history for deadline policies and configs
      description: >
        Signatures are re-verified on every read. `allSignaturesValid` says the remaining records
        are genuine; `systemChain` says none are missing.
      security: [{ oidc: [admin] }]
      parameters:
        - $ref: '#/components/parameters/CycleIdQuery'
      responses:
        '200':
          description: Activation history
          content:
            application/json:
              schema:
                type: object
                required: [activations, allSignaturesValid, systemChain]
                properties:
                  activations:
                    type: array
                    items: { $ref: '#/components/schemas/ActivationRecord' }
                  allSignaturesValid: { type: boolean }
                  systemChain: { $ref: '#/components/schemas/ChainVerification' }
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://id.example.kr/.well-known/openid-configuration
    mutualTLS:
      type: mutualTLS
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
...
    CycleIdQuery:
      in: query
      name: cycleId
      required: true
      schema: { type: string, format: uuid }
    ConfigId:
      in: path
      name: configId
      required: true
      schema: { type: string, format: uuid }
    PolicyId:
      in: path
      name: policyId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    ConfigVersion:
      type: object
      required: [id, version, status, configHash, createdBy, approvedBy, activatedAt]
      properties:
        id: { type: string, format: uuid }
        version: { type: string }
        status: { type: string, enum: [DRAFT, APPROVED, ACTIVE, RETIRED] }
        configHash: { type: string }
        createdBy: { type: string }
        approvedBy:
          type: array
          maxItems: 2
          items: { type: string }
        activatedAt: { type: [string, 'null'], format: date-time }
    CreateConfigVersionRequest:
      type: object
      required: [cycleId, version, config]
      properties:
        cycleId: { type: string, format: uuid }
        version: { type: string, maxLength: 64 }
        config:
          type: object
          additionalProperties: true
          description: >
            Includes `forms` (JSON Schema per admission type code) and the optional `retention`
            section ({ <category>: { days } }), validated against the retention matrix (D-38).
    CreateDeadlinePolicyRequest:
      type: object
      required: [cycleId, version, mode, deadlineAt]
      properties:
        cycleId: { type: string, format: uuid }
        version: { type: string }
        mode:
          type: string
          enum: [FINALIZED_COMMIT_BEFORE_DEADLINE, REQUEST_RECEIVED_BEFORE_DEADLINE, PAYMENT_APPROVED_BEFORE_DEADLINE]
        deadlineAt: { type: string, format: date-time }
...
    ChainVerification:
      type: object
      required: [valid, checked]
      properties:
        valid: { type: boolean }
        checked: { type: integer, minimum: 0 }
        brokenAt: { type: string, description: First event id where the hash chain breaks }
...
    SignatureStatus:
      type: string
      description: Re-verified on every read; not a stored value.
      enum: [VALID, INVALID, UNKNOWN_KEY]
    ActivateRequest:
      type: object
      properties:
        activateAt:
          type: string
          format: date-time
          description: Scheduled activation. Omit to activate now (the DB clock stamps the time, D-31).
    ActivationRecord:
      type: object
      description: Signed, append-only record of one activation event (T-M3-15).
      required: [activationId, subjectType, subjectId, subjectVersion, kind, effectiveAt, operatorId,
        reason, decisionRef, supersedesVersion, content, keyId, payloadHash, recordedAt, signature]
      properties:
        activationId: { type: string, format: uuid }
        subjectType: { type: string, enum: [DEADLINE_POLICY, CONFIG_VERSION] }
        subjectId: { type: string, format: uuid }
        subjectVersion: { type: string }
        kind: { type: string, enum: [ACTIVATE, EXTEND, ROLLBACK] }
        effectiveAt: { type: string, format: date-time }
        operatorId: { type: string }
        reason: { type: [string, 'null'] }
        decisionRef: { type: [string, 'null'], description: Admissions office decision document number (extensions) }
        supersedesVersion: { type: [string, 'null'] }
        content: { type: object, additionalProperties: true }
        keyId: { type: string }
        payloadHash: { type: string }
        recordedAt: { type: string, format: date-time }
        signature: { $ref: '#/components/schemas/SignatureStatus' }
    ConfigChange:
      type: object
      required: [path, kind, risk, summary]
      properties:
        path: { type: string }
        kind: { type: string, enum: [ADDED, REMOVED, CHANGED] }
        risk: { type: string, enum: [INFO, WARN, DESTRUCTIVE] }
        summary: { type: string }
        before: {}
        after: {}
...
    ConfigDiff:
      type: object
      required: [changes, destructive, digest, identical]
      properties:
        changes:
          type: array
          items: { $ref: '#/components/schemas/ConfigChange' }
        destructive:
          type: array
          items: { $ref: '#/components/schemas/ConfigChange' }
        digest: { type: string, description: Send back as acknowledgedDiffDigest when approving }
        identical: { type: boolean, description: An identical config is refused on approval }
        warnings:
          type: array
          description: >
            Config linter warnings (§A5) — works, but may not be what was meant (unknown admission
            type code, field without title, legacy profile fields). Not part of the digest. Errors
            (a schema that does not compile, malformed document lists) refuse the draft instead.
          items: { type: string }
    DeadlinePolicy:
      type: object
      required: [policyId, version, mode, deadlineAt, status, approvedBy, activatedAt, policyHash,
        createdBy, createdAt, extension]
      properties:
        policyId: { type: string, format: uuid }
        version: { type: string }
        mode: { type: string }
        deadlineAt: { type: string, format: date-time }
        status:
          type: string
          enum: [DRAFT, APPROVED, ACTIVATED]
          x-kadmission-register: D-23
          description: >
            Lifecycle only. Which policy is in force is decided by activated_at order
            (scheduled activation), not by this field.
        approvedBy:
          type: array
          maxItems: 2
          items: { type: string }
        activatedAt: { type: [string, 'null'], format: date-time }
        policyHash: { type: string }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        extension:
          type: [object, 'null']
          required: [kind, extendsVersion, extendsDeadlineAt, reason, decisionRef]
          properties:
            kind: { type: string, const: EXTENSION }
            extendsVersion: { type: string }
            extendsDeadlineAt: { type: string, format: date-time }
            reason: { type: string }
            decisionRef: { type: string }
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        traceId: { type: string }
        serverTime: { type: string, format: date-time }
security:
  - oidc: []